Fixing errors when there are no images

+++ public/server/services/MapService.js
@@ -29,24 +29,26 @@ exports.getMap = function (lat, lng, sensorId, i) {
         var pm25 = '';
         var uploadsDir  = __dirname+'/../../../images';
         fs.readdir(uploadsDir, function(err, files) {
-            files.forEach(function(file, index) {
-                fs.stat(path.join(uploadsDir, file), function(err, stat) {
-                    var endTime, now;
-                    if (err) {
-                        return console.error(err);
-                    }
-                    now = new Date().getTime();
-                    endTime = new Date(stat.ctime).getTime() + 86400000;
-                    if (now > endTime) {
-                        return rimraf(path.join(uploadsDir, file), function(err) {
-                            if (err) {
-                                return console.error(err);
-                            }
-                            console.log('successfully deleted');
-                        });
-                    }
+            if(files !== undefined) {
+                files.forEach(function (file, index) {
+                    fs.stat(path.join(uploadsDir, file), function (err, stat) {
+                        var endTime, now;
+                        if (err) {
+                            return console.error(err);
+                        }
+                        now = new Date().getTime();
+                        endTime = new Date(stat.ctime).getTime() + 86400000;
+                        if (now > endTime) {
+                            return rimraf(path.join(uploadsDir, file), function (err) {
+                                if (err) {
+                                    return console.error(err);
+                                }
+                                console.log('successfully deleted');
+                            });
+                        }
+                    });
                 });
-            });
+            }
         });
         sensorService.getSensorData(sensorId).then(function(res){
